requirements.txt
Flask
keyboard
mss
numpy
opencv_python
Pillow
Pillow
Pillow
PyAutoGUI
pytesseract
ultralytics
Werkzeug
gunicorn
deep-translator
